Refactor op
Refactor the EulerCharacteristic Op to use a RAI
Change name to describe implementation
